slight reorganization
