+++ test/utils.ts
@@ -1,7 +1,7 @@
-import { BigNumber, Contract, Wallet } from "ethers";
+import Table from "cli-table";
+import { BigNumber, Wallet } from "ethers";
 
 import { IERC20 } from "../contract-types/IERC20";
-import { TransferType } from "./l1.test";
 
 export type Balances = { alice: BigNumber; bob: BigNumber };
 
@@ -21,4 +21,43 @@ export async function getBalances(
 
     return { alice, bob };
   }
+}
+
+export type TransferType = "ERC20" | "ETH";
+
+export type ClaimTicketsScenario = {
+  transferType: TransferType;
+  batchSize: number;
+  amountOfTickets: number;
+};
+
+export type ClaimTicketsScenarioGasUsage = ClaimTicketsScenario & {
+  totalGasUsed: BigNumber;
+};
+
+export function printScenarioGasUsage(
+  scenarios: ClaimTicketsScenarioGasUsage[]
+) {
+  const table = new Table({
+    head: [
+      "Transfer Type",
+      "Total Gas Used",
+      "Total Claims",
+      "Batch Size",
+      " Average Gas Per Claim",
+    ],
+  });
+  for (const scenario of scenarios) {
+    const averagePerClaim = scenario.totalGasUsed
+      .div(scenario.amountOfTickets)
+      .toNumber();
+    table.push([
+      scenario.transferType,
+      scenario.totalGasUsed,
+      scenario.amountOfTickets,
+      scenario.batchSize,
+      averagePerClaim,
+    ]);
+  }
+  console.log(table.toString());
 }
